+++ packages/web-sdk/src/modules/region-picker/constants/element-ids.ts
@@ -1,0 +1,2 @@
+export const LISTBOX_ID_PREFIX = 'tlx-region-listbox';
+export const OPTION_ID_PREFIX = 'tlx-region-option';